# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Block": {
        "description": "One detected region on a PDF page.\n\nAttributes:\n    id: Stable identifier ``p<page>-<t|i><index>`` \u2014 see module docstring.\n    kind: ``\"text\"`` or ``\"image\"``.\n    bbox: ``[x_min, y_min, x_max, y_max]`` in millimetres, top-left origin.\n    text_sample: First ~60 characters of the block's text (text blocks only).\n    char_count: Number of non-whitespace characters (text blocks only).",
        "properties": {
          "bbox": {
            "items": {
              "type": "number"
            },
            "title": "Bbox",
            "type": "array"
          },
          "char_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Char Count"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "text",
              "image"
            ],
            "title": "Kind",
            "type": "string"
          },
          "text_sample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Sample"
          }
        },
        "required": [
          "id",
          "kind",
          "bbox"
        ],
        "title": "Block",
        "type": "object"
      },
      "Body_analyze_document_document_analyze_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_analyze_document_document_analyze_post",
        "type": "object"
      },
...
      "DocumentAnalysis": {
        "description": "Container returned by :func:`extract_blocks`.",
        "properties": {
          "pages": {
            "items": {
              "$ref": "#/components/schemas/PageBlocks"
            },
            "title": "Pages",
            "type": "array"
          }
        },
        "required": [
          "pages"
        ],
        "title": "DocumentAnalysis",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LayerPlan": {
        "description": "Per-layer print settings \u2014 the unified pivot type.\n\nReplaces the old ``GenerateLayer`` Pydantic model and\n``LayerGeneration`` dataclass. Every field is optional except\n``layer_id``; defaults are filled in later by\n:func:`pen_plotter.application.plan_resolver.resolve_plan` using\nthe active machine profile.\n\nNote on ``optimize`` / ``simplify_tolerance_mm``: today the SVG\ninside the surrounding :class:`PrintPlan` is already optimized by\nthe frontend's ``optimize \u2192 preflight \u2192 generate`` pipeline, so\nthese fields are traceability data \u2014 they ride along into the\nresolved plan and the persisted snapshot, and they participate in\n``plan_hash`` so equal SVG + different optimize intent cannot\ncollide. A later refactor (lot L3+) will let the application\nservices drive optimization directly from these fields.",
        "properties": {
          "assigned_color_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned Color Hex"
          },
          "color_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color Label"
          },
          "drawing_speed_mm_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drawing Speed Mm S"
          },
          "layer_id": {
            "title": "Layer Id",
            "type": "string"
          },
          "optimize": {
            "default": true,
            "title": "Optimize",
            "type": "boolean"
          },
          "pause_before": {
            "default": "auto",
            "enum": [
              "auto",
              "always",
              "never"
            ],
            "title": "Pause Before",
            "type": "string"
          },
          "simplify_tolerance_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simplify Tolerance Mm"
          },
          "source_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Color"
          },
          "target_pen_slot": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Pen Slot"
          }
        },
        "required": [
          "layer_id"
        ],
        "title": "LayerPlan",
        "type": "object"
      },
...
      "PageBlocks": {
        "description": "All detected blocks on one PDF page.",
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "title": "Blocks",
            "type": "array"
          },
          "height_mm": {
            "title": "Height Mm",
            "type": "number"
          },
          "page_index": {
            "title": "Page Index",
            "type": "integer"
          },
          "width_mm": {
            "title": "Width Mm",
            "type": "number"
          }
        },
        "required": [
          "page_index",
          "width_mm",
          "height_mm",
          "blocks"
        ],
        "title": "PageBlocks",
        "type": "object"
      },
...
      "PlacementPlan": {
        "description": "Sheet placement inside the machine workspace.\n\nMirrors :class:`pen_plotter.models.Placement` \u2014 kept as a separate\ndomain model so the pivot is self-contained and doesn't pull legacy\ntypes into the new package boundary.",
        "properties": {
          "offset_x_mm": {
            "default": 0.0,
            "title": "Offset X Mm",
            "type": "number"
          },
          "offset_y_mm": {
            "default": 0.0,
            "title": "Offset Y Mm",
            "type": "number"
          },
          "sheet_height_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Sheet Height Mm",
            "type": "number"
          },
          "sheet_width_mm": {
            "exclusiveMinimum": 0.0,
            "title": "Sheet Width Mm",
            "type": "number"
          }
        },
        "required": [
          "sheet_width_mm",
          "sheet_height_mm"
        ],
        "title": "PlacementPlan",
        "type": "object"
      },
      "PlanMetadata": {
        "description": "Provenance information attached to a plan.\n\nThe client may set ``client_version`` so a stored snapshot can later\nbe correlated with the UI release that produced it.",
        "properties": {
          "client_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Version"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "title": "PlanMetadata",
        "type": "object"
      },
...
      "PreflightReport": {
        "description": "Pre-run safety and estimation checks for a placed drawing.",
        "properties": {
          "drawing_length_mm": {
            "title": "Drawing Length Mm",
            "type": "number"
          },
          "estimated_seconds": {
            "title": "Estimated Seconds",
            "type": "number"
          },
          "height_mm": {
            "title": "Height Mm",
            "type": "number"
          },
          "layer_count": {
            "title": "Layer Count",
            "type": "integer"
          },
          "missing_pen_slots": {
            "items": {
              "type": "integer"
            },
            "title": "Missing Pen Slots",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "path_count": {
            "title": "Path Count",
            "type": "integer"
          },
          "pen_changes": {
            "title": "Pen Changes",
            "type": "integer"
          },
          "plan_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Hash"
          },
          "scale": {
            "title": "Scale",
            "type": "number"
          },
          "travel_length_mm": {
            "title": "Travel Length Mm",
            "type": "number"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          },
          "width_mm": {
            "title": "Width Mm",
            "type": "number"
          },
          "within_bounds": {
            "title": "Within Bounds",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "within_bounds",
          "width_mm",
          "height_mm",
          "scale",
          "drawing_length_mm",
          "travel_length_mm",
          "estimated_seconds",
          "pen_changes",
          "layer_count",
          "path_count"
        ],
        "title": "PreflightReport",
        "type": "object"
      },
      "PreflightRequest": {
        "description": "Wire model for ``POST /preflight`` \u2014 same shape as a print plan.",
        "properties": {
          "layers": {
            "items": {
              "$ref": "#/components/schemas/LayerPlan"
            },
            "title": "Layers",
            "type": "array"
          },
          "library_file_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Library File Id"
          },
          "margin_mm": {
            "default": 10.0,
            "title": "Margin Mm",
            "type": "number"
          },
          "metadata": {
            "$ref": "#/components/schemas/PlanMetadata"
          },
          "placement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlacementPlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "profile_name": {
            "title": "Profile Name",
            "type": "string"
          },
          "scale_mode": {
            "default": "fit",
            "enum": [
              "fit",
              "actual"
            ],
            "title": "Scale Mode",
            "type": "string"
          },
          "source_mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Mime"
          },
          "svg": {
            "title": "Svg",
            "type": "string"
          },
          "typography": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TypographyPlan"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "svg",
          "profile_name"
        ],
        "title": "PreflightRequest",
        "type": "object"
      },
      "PreflightSvgRequest": {
        "description": "Lightweight wire model for ``POST /preflight/svg``.\n\nAccepts just an already-rendered SVG + a profile name; skips the\nfull :class:`PrintPlan` ceremony. Used by the Compare drawer to\ncompute per-candidate metrics (drawing length, travel length,\nestimated time, pen-change count) without building a fake plan\naround each variant.",
        "properties": {
          "profile_name": {
            "title": "Profile Name",
            "type": "string"
          },
          "svg": {
            "title": "Svg",
            "type": "string"
          }
        },
        "required": [
          "svg",
          "profile_name"
        ],
        "title": "PreflightSvgRequest",
        "type": "object"
      },
...
      "TypographyPlan": {
        "description": "Typography settings for text sources, carried through the pivot.\n\nMirrors :class:`pen_plotter.typography.TypographyOptions` but lives\nin the domain layer (no Pillow / hershey-fonts dependency). The\nfields are forwarded into the resolved snapshot and the\n``plan_hash`` so two plans differing only in font / size / weight\ncan no longer collide.\n\nToday this rides as traceability data: the actual text\u2192SVG render\nstill happens at ``/upload`` time (or via ``_reprocess_existing``\nwhen the operator changes a setting and the frontend re-uploads).\nA later lot can let the application services re-render text\ndirectly from this plan; the data is ready for that day.",
        "properties": {
          "alignment": {
            "default": "left",
            "enum": [
              "left",
              "center",
              "right",
              "justify"
            ],
            "title": "Alignment",
            "type": "string"
          },
          "bold": {
            "default": false,
            "title": "Bold",
            "type": "boolean"
          },
          "font": {
            "default": "futural",
            "title": "Font",
            "type": "string"
          },
          "font_size_mm": {
            "default": 10.0,
            "exclusiveMinimum": 0.0,
            "maximum": 200.0,
            "title": "Font Size Mm",
            "type": "number"
          },
          "italic": {
            "default": false,
            "title": "Italic",
            "type": "boolean"
          },
          "letter_spacing_mm": {
            "default": 0.0,
            "maximum": 50.0,
            "minimum": -10.0,
            "title": "Letter Spacing Mm",
            "type": "number"
          },
          "line_spacing": {
            "default": 1.5,
            "exclusiveMinimum": 0.0,
            "title": "Line Spacing",
            "type": "number"
          },
          "margin_mm": {
            "default": 15.0,
            "minimum": 0.0,
            "title": "Margin Mm",
            "type": "number"
          },
          "page_height_mm": {
            "default": 297.0,
            "exclusiveMinimum": 0.0,
            "title": "Page Height Mm",
            "type": "number"
          },
          "page_width_mm": {
            "default": 210.0,
            "exclusiveMinimum": 0.0,
            "title": "Page Width Mm",
            "type": "number"
          },
          "stroke_width_mm": {
            "default": 0.3,
            "exclusiveMinimum": 0.0,
            "title": "Stroke Width Mm",
            "type": "number"
          }
        },
        "title": "TypographyPlan",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OmniPlot",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/document/analyze": {
      "post": {
        "description": "Extract text and image blocks for every page of a PDF.\n\nArgs:\n    file: A PDF upload; other MIME types are rejected with 415.\n\nReturns:\n    A :class:`DocumentAnalysis` listing the blocks of each page.\n\nRaises:\n    HTTPException: 415 when the upload is not a PDF; 413 if the\n        file exceeds the analysis size limit; 422 if PyMuPDF cannot\n        parse the document.",
        "operationId": "analyze_document_document_analyze_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_document_document_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAnalysis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Document"
      }
    },
...
    "/preflight": {
      "post": {
        "description": "Validate and estimate a placed drawing before generating it.\n\nThe response is a :class:`PreflightReport` with ``plan_hash`` set so\nthe frontend can verify it later matches the ``/generate`` hash.\n\nRaises:\n    HTTPException: 404 if the profile is unknown; 400 if the SVG\n        cannot be parsed or the plan fails business validation.",
        "operationId": "preflight_preflight_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight"
      }
    },
    "/preflight/svg": {
      "post": {
        "description": "Estimate metrics for a standalone SVG against a profile.\n\nLightweight variant that skips plan resolution / TypographyPlan\nre-rendering / placement composition. Returns the same\n:class:`PreflightReport` shape as the full ``/preflight`` so\nconsumers (e.g. the Compare drawer) can read\n``drawing_length_mm``, ``travel_length_mm``,\n``estimated_seconds`` and ``pen_changes`` straight from the\nreport.\n\nRaises:\n    HTTPException: 404 if the profile is unknown; 400 if the SVG\n        cannot be parsed.",
        "operationId": "preflight_svg_preflight_svg_post",
        "parameters": [
          {
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreflightSvgRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreflightReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preflight Svg"
      }
    },
...
  }
}